Improve EditableDashboardKpi loading
JIRA: RAIL-4158
